requirements.txt
# Requirements for the SmartSeed Backend
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==24.2.0
CacheControl==0.14.4
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
click==8.3.1
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.3
cycler==0.12.1
discord==2.3.2
discord.py==2.4.0
docutils==0.21.2
fastapi==0.128.0
firebase_admin==7.1.0
fonttools==4.55.3
frozenlist==1.5.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
keyring==25.5.0
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.9.4
mdurl==0.1.2
more-itertools==10.5.0
msgpack==1.1.2
multidict==6.1.0
nh3==0.2.18
numpy==2.1.3
packaging==24.2
pillow==11.0.0
pkginfo==1.10.0
propcache==0.2.0
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.18.0
PyJWT==2.10.1
pyloco==0.0.139
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pywin32-ctypes==0.2.3
readme_renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.9.4
rsa==4.9.1
SimpleWebSocketServer==0.1.2
six==1.16.0
starlette==0.50.0
twine==6.0.1
typing==3.7.4.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.3
ushlex==0.99.1
uvicorn==0.40.0
websocket-client==1.8.0
yarl==1.18.0
zipp==3.21.0
earthengine-api==0.1.389
